2.2.0 [...]  - New "bandLimit" settings  - Better defaults for advanced humidity. Slightly faster, much more accurate;  - noiseFactor is now specified as a percentage of bandSize by default  - Preliminary support for advanced humidity under Custom Biome Parameters (and maybe other mods that change river/ocean biome parameters)  - Fix potential integer overflow during benchmark;  - Fix for dedicated servers (and general improvements to world creation/loading process, esp. for multithreaded environments)

+++ common/src/main/java/github/cosmicdan/temperaturebands/TbUtils.java
@@ -217,8 +217,8 @@ public class TbUtils {
             if (batchesDone.get() >= batchesTotal.get()) {
                 benchmarkActive = false;
                 double timeMillis = Duration.between(benchmarkStart.get(), Instant.now()).abs().toMillis();
-                int cacheHitCount = BiomeProximityGenerator.benchmarkSampleCacheHitCount.get();
-                int cacheTotalCount = BiomeProximityGenerator.benchmarkSampleTotalCount.get();
+                long cacheHitCount = BiomeProximityGenerator.benchmarkSampleCacheHitCount.get();
+                long cacheTotalCount = BiomeProximityGenerator.benchmarkSampleTotalCount.get();
                 TemperatureBands.LOGGER.info("WorldPreview benchmark finished. {} chunks via {} batches finished in {} seconds. Average generation speed of {} chunks per second. Sampler cache hit-rate was {}% ({} over {})." ,
                         chunksTotal,
                         batchesTotal,
